# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KISS Notebook API",
    "version": "1.0.0",
    "description": "REST API for collaborative Jupyter-like notebook platform.\n\nAll successful responses are wrapped in `{\"data\": ...}`; errors in `{\"error\": \"message\"}`.\n\n**Auth flow in Swagger UI:**\n1. `POST /api/v1/auth/login` — response sets `session_id` and `csrf_token` cookies.\n2. Copy the `csrf_token` cookie value.\n3. Click **Authorize**: set `cookieAuth` = `session_id` value, set `csrfToken` = `csrf_token` value.\n4. All subsequent mutating requests will include `X-CSRF-Token` header automatically."
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Local development"
    }
  ],
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataUserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login and create session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Sets `session_id` and `csrf_token` cookies.",
            "headers": {
              "Set-Cookie": {
                "schema": { "type": "string" },
                "description": "session_id and csrf_token cookies"
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataUserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout and destroy session",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get current user",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataUserResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": ["admin"],
        "summary": "List all users (admin only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataAdminUserListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/admin/users/{id}/ban": {
      "post": {
        "tags": ["admin"],
        "summary": "Ban a user (admin only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/UserPathID" }],
        "responses": {
          "200": {
            "description": "User banned (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/admin/users/{id}/unban": {
      "post": {
        "tags": ["admin"],
        "summary": "Unban a user (admin only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/UserPathID" }],
        "responses": {
          "200": {
            "description": "User unbanned (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/admin/notebooks": {
      "get": {
        "tags": ["admin"],
        "summary": "List all notebooks (admin only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of notebooks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataAdminNotebookListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/admin/notebooks/{id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Delete any notebook (admin only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NotebookID" }],
        "responses": {
          "200": {
            "description": "Notebook deleted (returns `{\"data\": null}`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataNullResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "tags": ["admin"],
        "summary": "Get aggregate platform stats (admin only)",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Aggregate stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DataAdminStatsResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "parameters": {
      "NotebookID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
...
      "UserPathID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "Resource conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DataNullResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "null", "nullable": true }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "avatar_url": { "type": "string" },
          "status": { "type": "string" },
          "description": { "type": "string" },
          "is_admin": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "DataUserResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/UserResponse" }
        }
      },
...
      "AdminUserListResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          },
          "total": { "type": "integer" }
        }
      },
      "DataAdminUserListResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminUserListResponse" }
        }
      },
      "AdminNotebookItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "owner_id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "is_public": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "AdminNotebookListResponse": {
        "type": "object",
        "properties": {
          "notebooks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminNotebookItem" }
          },
          "total": { "type": "integer" }
        }
      },
      "DataAdminNotebookListResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminNotebookListResponse" }
        }
      },
      "AdminStatsResponse": {
        "type": "object",
        "properties": {
          "total_users": { "type": "integer" },
          "total_sessions": { "type": "integer" },
          "dau": { "type": "integer", "description": "Daily active users" },
          "mau": { "type": "integer", "description": "Monthly active users" }
        }
      },
      "DataAdminStatsResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminStatsResponse" }
        }
      }
    }
  }
}